Moved dpos utilities and updated file imports
