+++ packages/block/store/action.test.js
@@ -1,6 +1,6 @@
 import networks from '@network/configuration/networks';
 import { getBlocks } from '@block/utilities/api';
-import { getForgers, getDelegates } from '@dpos/utilities/api';
+import { getForgers, getDelegates } from '@dpos/validator/utiles/api';
 import forger from '@fixtures/forgers';
 import delegate from '@fixtures/delegates';
 import blocks from '@tests/constants/blocks';
@@ -10,7 +10,7 @@ import { olderBlocksRetrieved, forgersRetrieved } from './action';
 jest.mock('@block/utilities/api', () => ({
   getBlocks: jest.fn(),
 }));
-jest.mock('@dpos/utilities/api', () => ({
+jest.mock('@dpos/validator/utiles/api', () => ({
   getForgers: jest.fn(),
   getDelegates: jest.fn(),
 }));
