+++ setup/.eslintrc
@@ -35,7 +35,7 @@
           ["@block", "./packages/block/"],
           ["@bookmark", "./packages/bookmark/"],
           ["@common", "./packages/common/"],
-          ["@dpos", "./packages/dpos/"],
+          ["@dpos", "./src/modules/dpos/"],
           ["@network", "./packages/network/"],
           ["@settings", "./packages/settings/"],
           ["@token", "./packages/token/"],
